requirements.txt
pyyaml>=6.0
networkx>=3.2
neo4j>=5.15.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
groq>=0.4.1
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
streamlit>=1.28.0
plotly>=5.17.0
websockets>=12.0
